+++ src/main.tsx
@@ -1,7 +1,7 @@
+import { reportWebVitals } from '@/shared/components/WebVitalsMonitor/reportWebVitals'
 import React from 'react'
 import ReactDOM from 'react-dom/client'
 import App from './App.tsx'
-import { reportWebVitals } from './reportWebVitals.ts'
 import './shared/model/styles/globals.css'
 
 ReactDOM.createRoot(document.getElementById('root')!).render(
